rfid_python_ui/pyproject.toml
[project]
name = "rfid-python-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "nicegui>=3.0.4",
    "numpy>=2.2.6",
    "pymysql>=1.1.2",
    "pyserial>=3.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tk>=0.1.0",
]
